Add files via upload
In this project, "Sentiment Analysis using NLP on IMDB Movie Reviews" you will understand the sentiments and opinions expressed in a diverse collection of movie reviews from the IMDB website.  The main goal was to harness the power of NLP techniques to automatically classify movie reviews as either positive or negative based on the text content.